TG-60: redux implementation completed. Error modal bug WIP

+++ src/Components/Admins/List/List.jsx
@@ -4,41 +4,18 @@ import Row from '../../Shared/Row/Row';
 import { useDispatch, useSelector } from 'react-redux/es/exports';
 import { deleteAdmins } from '../../../redux/admins/thunks';
 
-const List = ({
-  // setShowModal,
-  // setChildrenModal,
-  setShowForm,
-  setPreviousAdmin,
-  setMethod
-}) => {
+const List = ({ setShowForm, setPreviousAdmin }) => {
   const dispatch = useDispatch();
 
   const admins = useSelector((state) => state.admins.list);
 
-  // const handleDelete = async (_id) => {
-  //   dispatch(deleteAdminsPending());
-  //   if (confirm('Are you sure you want to remove the Admin?')) {
-  //     try {
-  //       await fetch(`${process.env.REACT_APP_API_URL}/admins/${_id}`, {
-  //         method: 'DELETE'
-  //       });
-  //       setShowModal(true);
-  //       dispatch(deleteAdminsSuccess(_id));
-  //       setChildrenModal('Admin deleted successfully');
-  //     } catch (error) {
-  //       setShowModal(true);
-  //       setChildrenModal(error.msg);
-  //       dispatch(deleteAdminsError(error));
-  //     }
-  //   }
-  // };
-
   const handleDelete = async (_id) => {
-    dispatch(deleteAdmins(_id));
+    if (confirm('Are you sure you want to remove the Admin?')) {
+      dispatch(deleteAdmins(_id));
+    }
   };
 
   const handleEdit = (admin) => {
-    setMethod('PUT');
     setPreviousAdmin(admin);
     setShowForm(true);
   };
